clean

+++ srcs/ft_floor.c
@@ -6,14 +6,14 @@
 /*   By: mbos <mbos@student.le-101.fr>              +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2020/01/06 13:34:20 by mbos         #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/06 13:34:35 by mbos        ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/06 16:52:04 by mbos        ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-long long		ft_floor(long double nb)
+long long		ft_floor(float nb)
 {
 	long long	nbi;
